refactor: View 검증, Domain 검증 분리
