+++ src/test/java/racingcar/domain/RacingCarTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.view.Validator;
 
 public class RacingCarTest {
 
@@ -40,4 +42,11 @@ public class RacingCarTest {
             .withMessageMatching("자동차 이름은 비어있을 수 없다.");
     }
 
+    @Test
+    @DisplayName("자동차의 이름이 5자가 초과될 경우 예외가 발생해야 한다.")
+    void checkRacingCarNameLength() {
+        assertThatExceptionOfType(IllegalArgumentException.class)
+            .isThrownBy(() -> new RacingCar("abcdef"))
+            .withMessageMatching("자동차 이름은 5자를 초과할 수 없다.");
+    }
 }
